refactor: simplify code logic and add CallParticipantsBadge component
